refactor: move element operations to utils
